Need to use isset instead of property_exists
Seems there is a weird tension in the use of ArrayObject where you
can't use array_key_exists because it is still an object, but it can
have keys/values like one. Looking at a sample response, `next_cursor`
is a key at the same level as `resources`, so it seems `isset` is the
only option.

Logically I think that still works because if the value of
`$response['next_cursor']` is null, then this code should not execute.